digest: factor out `fixed`, `variable`, and `xof` modules
Each of these now defines two traits and a blanket impl, so `lib.rs` was
getting rather bloated.

This commit factors each of them into their own modules.
